refactor: Split UserServiceResponseHelpers test file to meet <500 line requirement
- Split UserServiceResponseHelpers.comprehensive.test.ts into 3 focused files:
  - UserServiceResponseHelpers.core.test.ts (182 lines)
  - UserServiceResponseHelpers.errorHandling.test.ts (233 lines)
  - UserServiceResponseHelpers.serialization.test.ts (219 lines)
- All test files now under 500 lines to meet code complexity standards
- Maintain 100% test coverage for UserServiceResponseHelpers module
- Fix ESLint issues (unused imports, missing newlines)
- All 236 UserService tests continue to pass

Relates to #106
